// frontend/src/api/schemas/profileUpdateRequest.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * URL Shortener API
 * RESTful API for URL shortening service. Provides endpoints for URL shortening, redirection, user management, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export interface ProfileUpdateRequest {
	/** @minLength 1 */
	name: string
	/** @minLength 1 */
	email: string
}